subtree updates
meta-openembedded: 9a0caf5b09..0e6c34f82c:
  Martin Jansa (1):
        ostree: prevent ostree-native depending on target virtual/kernel to provide kernel-module-overlay

  Nicolas Dechesne (1):
        imlib2: update SRC_URI

poky: eff78b3802..fd00d74f47:
  Alexandre Belloni (1):
        pseudo: Fix handling of absolute links

  Anuj Mittal (2):
        poky.conf: bump version for 3.4.4 release
        documentation: update for 3.4.4 release

  Bruce Ashfield (10):
        linux-yocto/5.10: update to v5.10.109
        lttng-modules: update to 2.13.1
        lttng-modules: support kernel 5.18+
        linux-yocto/5.10: features/security: Move x86_64 configs to separate file
        linux-yocto/5.10: update to v5.10.110
        linux-yocto/5.10: base: enable kernel crypto userspace API
        linux-yocto/5.10: update to v5.10.112
        linux-yocto: enable powerpc debug fragment
        linux-yocto/5.10: update to v5.10.113
        yocto-bsps: update to v5.10.113

  Chen Qi (1):
        cases/buildepoxy.py: fix typo

  Davide Gardenal (1):
        rootfs-postcommands: fix symlinks where link and output path are equal

  Dmitry Baryshkov (2):
        linux-firmware: correct license for ar3k firmware
        linux-firmware: upgrade 20220411 -> 20220509

  Felix Moessbauer (1):
        wic/plugins/rootfs: Fix permissions when splitting rootfs folders across partitions

  Joe Slater (1):
        unzip: fix CVE-2021-4217

  Khem Raj (2):
        busybox: Use base_bindir instead of hardcoding /bin path
        linux-yocto: Enable powerpc-debug fragment for ppc64 LE

  Konrad Weihmann (1):
        gmp: add missing COPYINGv3

  Martin Jansa (1):
        license_image.bbclass: close package.manifest file

  Max Krummenacher (2):
        perf: sort-pmuevents: don't drop elements
        perf: sort-pmuevents: allow for additional type qualifiers and storage class

  Michael Opdenacker (1):
        adding missing space in appends

  Ovidiu Panait (1):
        openssl: upgrade 1.1.1l -> 1.1.1n

  Paul Gortmaker (1):
        install/devshell: Introduce git intercept script due to fakeroot issues

  Peter Kjellerstedt (1):
        u-boot: Inherit pkgconfig

  Portia (1):
        volatile-binds: Change DefaultDependencies from false to no

  Rahul Kumar (1):
        neard: Switch SRC_URI to git repo

  Ralph Siemsen (1):
        xz: fix CVE-2022-1271

  Ranjitsinh Rathod (1):
        openssl: Minor security upgrade 1.1.1n to 1.1.1o

  Richard Purdie (14):
        scripts/runqemu: Fix memory limits for qemux86-64
        vim: Upgrade 8.2.4524 -> 8.2.4681
        uninative: Upgrade to 3.6 with gcc 12 support
        tiff: Add marker for CVE-2022-1056 being fixed
        externalsrc/devtool: Fix to work with fixed export funcition flags handling
        libxshmfence: Correct LICENSE to HPND
        alsa-tools: Ensure we install correctly
        shadow-native: Simplify and fix syslog disable patch
        build-appliance-image: Update to honister head revision
        base: Avoid circular references to our own scripts
        base: Drop git intercept
        scripts: Make git intercept global
        scripts/git: Ensure we don't have circular references
        vim: Upgrade 8.2.4681 -> 8.2.4912

  Ross Burton (4):
        bitbake: knotty: display active tasks when printing keepAlive() message
        bitbake: knotty: reduce keep-alive timeout from 5000s (83 minutes) to 10 minutes
        bitbake.conf: mark all directories as safe for git to read
        oeqa/selftest: add test for git working correctly inside pseudo

  wangmy (3):
        linux-firmware: upgrade 20220310 -> 20220411
        lttng-modules: upgrade 2.13.1 -> 2.13.2
        lttng-modules: upgrade 2.13.2 -> 2.13.3

  zhengruoqin (1):
        wireless-regdb: upgrade 2022.02.18 -> 2022.04.08

+++ poky/meta/classes/rootfs-postcommands.bbclass
@@ -267,9 +267,10 @@ python write_image_manifest () {
 
     if os.path.exists(manifest_name) and link_name:
         manifest_link = deploy_dir + "/" + link_name + ".manifest"
-        if os.path.lexists(manifest_link):
-            os.remove(manifest_link)
-        os.symlink(os.path.basename(manifest_name), manifest_link)
+        if manifest_link != manifest_name:
+            if os.path.lexists(manifest_link):
+                os.remove(manifest_link)
+            os.symlink(os.path.basename(manifest_name), manifest_link)
 }
 
 # Can be used to create /etc/timestamp during image construction to give a reasonably
@@ -339,9 +340,10 @@ python write_image_test_data() {
 
     if os.path.exists(testdata_name) and link_name:
         testdata_link = os.path.join(deploy_dir, "%s.testdata.json" % link_name)
-        if os.path.lexists(testdata_link):
-            os.remove(testdata_link)
-        os.symlink(os.path.basename(testdata_name), testdata_link)
+        if testdata_link != testdata_name:
+            if os.path.lexists(testdata_link):
+                os.remove(testdata_link)
+            os.symlink(os.path.basename(testdata_name), testdata_link)
 }
 write_image_test_data[vardepsexclude] += "TOPDIR"
